Put IEA15 test case in test_rotor.cpp
